requirements.txt
streamlit>=1.36
python-dotenv>=1.0.1
pillow>=10.3
pymupdf>=1.24.7
chromadb>=0.5.5
openai>=1.40.0
tiktoken>=0.7.0

# Adobe PDF Services SDK
pdfservices-sdk>=3.0.0
supabase>=2.6.0

# 오디오 녹음 컴포넌트
audio_recorder_streamlit
pydub